terminando de mexer na dashboard do administrador e aplicando permisseos no sistema

+++ src/app/Http/Controllers/DashboardController.php
@@ -2,7 +2,10 @@
 
 namespace App\Http\Controllers;
 
+use App\Room;
 use App\User;
+use App\Category;
+use App\Modality;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Redirect;
@@ -15,11 +18,19 @@ class DashboardController extends Controller
     public function __construct()
     {
         $this->middleware('auth');
+        $this->middleware('role.admin')->except('editUser');
+
     }
 
     public function show(Request $request ){
+        $allRooms = Room::paginate(6);
         $users = User::paginate(6);
-        return view('dashboard.panel' , compact('users') );
+        $usersTotal = User::count(); 
+        $categorysTotal =  Category::count();
+        $modalityTotal= Modality::count();
+        $roomsTotal = Room::count();
+        //dd($users);
+        return view('dashboard.panel' , compact('users', 'usersTotal' , "categorysTotal" , "modalityTotal" , "roomsTotal",'allRooms'));
     }
 
     public function  editUser($slugName)
@@ -30,8 +41,10 @@ class DashboardController extends Controller
 
     public function deleteUser($slugName)
     {
+        if($slugName === Auth::user()->slug) return  view('errors.403');
 
         if( User::where('slug' , $slugName)->delete() )  return Redirect::route('dashboard')->with("message","Usuario Deleteado Com Sucesso");
+        
 
         return  view('errors.404');
